+++ packages/pluggableWidgets/gallery-native/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.0.1] - 2025-10-17
+
 ### Fixed
 
 -  We've fixed an issue where the Gallery widget would not properly load more items when scrolling down quickly.